EmailAutomation/requirements.txt
O365==2.0.22
pandas==2.0.3
sqlalchemy==2.0.7
python-dotenv==1.0.0
requests==2.31.0
